Add files via upload
The logo has the water mark because Adobe Express refused to allow me to publish it unless I used their premium service. There is a free trial available, but starting one of those without having a plan in mind for it or an immediate desire to use the paid program (which requires a plan in mind for it) makes my skin crawl in the same way, albeit less, than paying for textbooks that I can find free otherwise.